Fixing Typo in VisualConceptTagImportHandler, removing lsc-specific configs
